+++ playwright.config.ts
@@ -1,40 +1,47 @@
-import { devices, type PlaywrightTestConfig } from '@playwright/test'
+import { defineConfig, devices } from '@playwright/test'
 
-const config: PlaywrightTestConfig = {
+export default defineConfig({
   testDir: './tests/e2e',
-  webServer: {
-    command: 'pnpm dev',
-    url: 'http://localhost:3000/',
-    timeout: 120 * 1000,
-    reuseExistingServer: !process.env.CI,
+  timeout: 30 * 1000,
+  expect: {
+    timeout: 5000,
   },
+  forbidOnly: !!process.env.CI,
+  retries: process.env.CI ? 2 : 0,
+  workers: process.env.CI ? 1 : 4,
   use: {
-    baseURL: 'http://localhost:3000',
+    actionTimeout: 0,
+    baseURL: 'http://localhost:3001',
     trace: 'on-first-retry',
+  },
+  webServer: {
+    command: 'pnpm preview',
+    port: 3001,
+    reuseExistingServer: !process.env.CI,
+    stdout: 'ignore',
+    stderr: 'pipe',
   },
   projects: [
     {
-      name: 'iPhone 6',
+      name: 'webkit',
       use: {
         browserName: 'webkit',
-        ...devices['iPhone 6'],
+        ...devices['iPhone SE'],
       },
     },
     {
-      name: 'Macbook 11',
+      name: 'firefox',
       use: {
         browserName: 'firefox',
         ...devices['Macbook 11'],
       },
     },
     {
-      name: 'Desktop',
+      name: 'chromium',
       use: {
         browserName: 'chromium',
         ...devices['Macbook Pro'],
       },
     },
   ],
-}
-
-export default config
+})